// src/common/enums/response-messages.enum.ts
export enum ResponseMessages {
  SUCCESS = 'SUCCESS',
  BAD_GATEWAY = 'BAD_GATEWAY',
  NOT_FOUND = 'NOT_FOUND',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  UNAUTHORIZATION = 'UNAUTHORIZATION',
  FORBIDDEN = 'FORBIDDEN',
  BAD_REQUEST = 'BAD_REQUEST',
  INVALID_CARD_PROVIDER_NAME = 'INVALID_CARD_PROVIDER_NAME',
  INVALID_PAYMENT_PROVIDER_NAME = 'INVALID_PAYMENT_PROVIDER_NAME',
  INVALID_3DS_PAYMENT_PROVIDER_NAME = 'INVALID_3DS_PAYMENT_PROVIDER_NAME',
  INVALID_USER_PROVIDER_NAME = 'INVALID_USER_PROVIDER_NAME',
}